Fixed the memory breaking screen transitions
Kind of cheating by saying the standards let me do this... Which isn't
really cheating then. I can neatly pass shared pointers since the
Transitioner is part owner of the Scene. So that fixes that.

Also spotted a bug with the Fade transition but left a comment and can
fix that later.